# swagger.json
{"openapi":"3.0.0","paths":{"/api/companies":{"post":{"operationId":"CompaniesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":"Company created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyResponseDto"}}}}},"summary":"Create company","tags":["companies"]},"get":{"operationId":"CompaniesController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"document","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Companies retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaginatedCompaniesResponseDto"}}}}}},"summary":"List all companies","tags":["companies"]}},"/api/companies/{id}":{"get":{"operationId":"CompaniesController_findById","parameters":[{"name":"id","required":true,"in":"path","description":"Company ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Company retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyResponseDto"}}}},"404":{"description":"Company not found"}},"summary":"Get company by ID","tags":["companies"]},"put":{"operationId":"CompaniesController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Company ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":"Company updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyResponseDto"}}}},"404":{"description":"Company not found"}},"summary":"Update company","tags":["companies"]},"delete":{"operationId":"CompaniesController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"Company ID","schema":{"type":"string"}}],"responses":{"204":{"description":"Company deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyResponseDto"}}}},"404":{"description":"Company not found"}},"summary":"Delete company","tags":["companies"]}},"/api/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"description":"User credentials","content":{"application/json":{"schema":{"properties":{"email":{"type":"string","description":"Valid email"},"password":{"type":"string","description":"Valid password"}}}}}},"responses":{"200":{"description":"Successfully authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Authenticate user","tags":["auth"]}},"/api/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"200":{"description":"Successfully registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Register user","tags":["auth"]}},"/api/campaigns":{"post":{"operationId":"CampaignsController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"phones-list":{"type":"string","format":"binary","description":"Text file containing phone numbers in format XX XXXXX-XXXX, one phone for line"},"name":{"type":"string","description":"Name of the campaign"},"message":{"type":"string","description":"Campaign message content"}},"required":["phones-list","name","message"]}}}},"responses":{"201":{"description":"Campaign created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponseDto"}}}},"400":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create campaign","tags":["campaigns"]},"get":{"operationId":"CampaignsController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string","enum":["PENDING","SENDED"]}}],"responses":{"200":{"description":"Campaigns retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaginatedCampaignsResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"List all campaigns","tags":["campaigns"]}},"/api/campaigns/{id}":{"get":{"operationId":"CampaignsController_findById","parameters":[{"name":"id","required":true,"in":"path","description":"Campaign ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Campaign retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponseDto"}}}},"404":{"description":"Campaign not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get campaign by ID","tags":["campaigns"]},"put":{"operationId":"CampaignsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Campaign ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCampaignDto"}}}},"responses":{"200":{"description":"Campaign updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponseDto"}}}},"404":{"description":"Campaign not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update campaign","tags":["campaigns"]},"delete":{"operationId":"CampaignsController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"Campaign ID","schema":{"type":"string"}}],"responses":{"204":{"description":"Campaign deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignResponseDto"}}}},"404":{"description":"Campaign not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Delete campaign","tags":["campaigns"]}}},"info":{"title":"Campaign Management API","description":"API for managing messages campaigns","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateCompanyDto":{"type":"object","properties":{"document":{"type":"string"},"name":{"type":"string"}},"required":["document","name"]},"CompanyResponseDto":{"type":"object","properties":{"id":{"type":"string"},"document":{"type":"string"},"name":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","document","name","created_at","updated_at"]},"PaginatedCompaniesResponseDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CompanyResponseDto"}},"page":{"type":"number"},"totalPages":{"type":"number"},"totalItems":{"type":"number"}},"required":["data","page","totalPages","totalItems"]},"UpdateCompanyDto":{"type":"object","properties":{"document":{"type":"string"},"name":{"type":"string"}}},"AuthResponseDto":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"ErrorResponseDto":{"type":"object","properties":{"statusCode":{"type":"number"},"timestamp":{"type":"string"},"path":{"type":"string"},"method":{"type":"string"},"message":{"type":"array","items":{"type":"string"}}},"required":["statusCode","timestamp","path","method","message"]},"RegisterUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"company_id":{"type":"string"}},"required":["email","password","company_id"]},"CampaignResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"company_id":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","name","status","company_id","created_at","updated_at"]},"PaginatedCampaignsResponseDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CampaignResponseDto"}},"page":{"type":"number"},"totalPages":{"type":"number"},"totalItems":{"type":"number"}},"required":["data","page","totalPages","totalItems"]},"UpdateCampaignDto":{"type":"object","properties":{"name":{"type":"string"}}}}}}